Only compute substitutions when necessary
Previously, since gsub was invoked with a hash, all substitutions would
be computed regardless of whether the relevant placeholder was found in
the prompt string.

This commit changes the invocation to use the block form of gsub. The
passed block will be executed once for every placeholder found in the
string. On each execution it will only calculate the relevant
substitution.