Cache repo info during prompt generation
When generating the prompt, the prompter and prompt color would
repeatedly ask the environment for information about the repo. This
resulted in many calls to external git commands, slowing down the prompt
generation.

This commit reduces the number of calls to external git commands to the
minimum possible to retain the previous behavior by:

1. Caching information within the prompter during calls to #prompt.
2. Inlining the prompt color #status_color method within the prompter to
   be able to make use of the cache.